add xcm test to post rt upgrade suite
