+++ .github/workflows/integration-tests.yml
@@ -103,7 +103,7 @@ jobs:
           ./integration-tests/scripts/deploy-zombienet.sh --test
 
   zombienet_zeitgeist_upgrade:
-    name: Zombienet Runtime Upgrade Zeitgeist
+    name: Zeitgeist Zombienet Post-Upgrade Tests
     runs-on: ubuntu-20.04
     needs: ["build_parachain"]
     steps:
@@ -132,14 +132,14 @@ jobs:
 
       - name: Create local folders
         run: |
-          mkdir -p target/release/wbuild/battery-station-runtime/
+          mkdir -p target/release/wbuild/zeitgeist-runtime/
           mkdir -p integration-tests/tmp
 
       - name: Download runtime
         uses: actions/download-artifact@v3.0.2
         with:
           name: runtimes
-          path: target/release/wbuild/battery-station-runtime/
+          path: target/release/wbuild/zeitgeist-runtime/
 
       - name: Download binary
         uses: actions/download-artifact@v3.0.2
@@ -159,7 +159,7 @@ jobs:
           pnpm exec moonwall test zombienet_zeitgeist_upgrade
 
   chopsticks_battery_station_upgrade:
-    name: Chopsticks Runtime Upgrade Battery Station
+    name: Battery Station Chopsticks Post-Upgrade Tests
     runs-on: ubuntu-20.04
     needs: ["build_parachain"]
     steps:
@@ -202,7 +202,7 @@ jobs:
         run: ls -R
         working-directory: target/
 
-      - name: Test battery station runtime upgrade using Chopsticks
+      - name: Battery Station post-upgrade tests using Chopsticks
         run: |
           cd integration-tests
           pnpm exec moonwall test chopsticks_battery_station_upgrade
@@ -210,7 +210,7 @@ jobs:
           cat tmp/node_logs/*.log
 
   chopsticks_zeitgeist_upgrade:
-    name: Chopsticks Runtime Upgrade Zeitgeist
+    name: Zeitgeist Chopsticks Post-Upgrade Tests
     runs-on: ubuntu-20.04
     needs: ["build_parachain"]
     steps:
@@ -253,50 +253,9 @@ jobs:
         run: ls -R
         working-directory: target/
 
-      - name: Test zeitgeist runtime upgrade using Chopsticks
+      - name: Zeitgeist post-upgrade tests using Chopsticks
         run: |
           cd integration-tests
           pnpm exec moonwall test chopsticks_zeitgeist_upgrade
           ls -R tmp/node_logs/
           cat tmp/node_logs/*.log
-
-  chopsticks_zeitgeist_hydradx_xcm_tests:
-    name: Chopsticks Zeitgeist HydraDX XCM Tests
-    runs-on: ubuntu-20.04
-    needs: ["build_parachain"]
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v2
-
-      - name: Install build tools
-        run: ./scripts/init.sh
-
-      - uses: pnpm/action-setup@v2
-        with:
-          version: 8
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 20.x
-          cache: "pnpm"
-          cache-dependency-path: "./integration-tests/pnpm-lock.yaml"
-
-      - name: Install pnpm packages
-        run: |
-          cd integration-tests
-          pnpm install
-
-      - name: Cache Dependencies
-        uses: Swatinem/rust-cache@v1
-
-      - name: Create local folders
-        run: |
-          mkdir -p integration-tests/tmp
-          mkdir -p integration-tests/tmp/node_logs
-
-      - name: Test zeitgeist xcm tests using Chopsticks
-        run: |
-          cd integration-tests
-          pnpm exec moonwall test chopsticks_zeitgeist_hydra_xcm_tests
-          ls -R tmp/node_logs/
-          cat tmp/node_logs/*.log
-
